added output resolution rendering
