+++ osd/const.py
@@ -1,3 +1,5 @@
+from .utils.ro_cls import read_only_class
+
 DEFAULT_SECTION = 'DEFAULT'
 
 SD_TILE_WIDTH = 12 * 3
@@ -23,3 +25,24 @@ FW_INAV = 1
 FW_BETAFL = 2
 FW_ARDU = 3
 
+@read_only_class
+class ArduParams:
+    LAT_CHAR_CODE: int = 167
+    LON_CHAR_CODE: int = 166
+    ALT_CHAR_CODE: int = 177
+    HOME_CHAR_CODE: int = 191
+
+    ALT_LEN: int = 4
+    GPS_LEN: int = 12
+    HOME_LEN: int = 6
+
+@read_only_class
+class InavParams:
+    LAT_CHAR_CODE: int = 3
+    LON_CHAR_CODE: int = 4
+    ALT_CHAR_CODE: int = 118
+    HOME_CHAR_CODE: int = 16
+
+    ALT_LEN: int = 4
+    GPS_LEN: int = 9
+    HOME_LEN: int = 5
